+++ packages/contracts-rfq/contracts/interfaces/IAdminV2Errors.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.4;
+
+interface IAdminV2Errors {
+    error CancelDelayBelowMin();
+    error FeeRateAboveMax();
+}