+++ packages/rest-api/src/routes/index.ts
@@ -11,6 +11,8 @@ import destinationTxRoute from './destinationTxRoute'
 import tokenListRoute from './tokenListRoute'
 import destinationTokensRoute from './destinationTokensRoute'
 import bridgeLimitsRoute from './bridgeLimitsRoute'
+import chainIconRoute from './chainIconRoute'
+import addressIconRoute from './addressIconRoute'
 
 const router: express.Router = express.Router()
 
@@ -25,5 +27,7 @@ router.use('/bridgeTxStatus', bridgeTxStatusRoute)
 router.use('/destinationTx', destinationTxRoute)
 router.use('/tokenList', tokenListRoute)
 router.use('/destinationTokens', destinationTokensRoute)
+router.use('/chainIcon', chainIconRoute)
+router.use('/tokenIcon', addressIconRoute)
 
 export default router